[OrderAdapter] Move to data_adapter
