+++ octobot_trading/data_adapters/order_adapter.pxd
