+++ octobot_trading/traders/trader.py
@@ -21,7 +21,7 @@ from octobot_trading.constants import REAL_TRADER_STR, CONFIG_TRADER_RISK, CONFI
 from octobot_trading.data.order import Order
 from octobot_trading.data.portfolio import Portfolio
 from octobot_trading.enums import OrderStatus, TraderOrderType
-from octobot_trading.orders.order_adapter import check_and_adapt_order_details_if_necessary
+from octobot_trading.data_adapters.order_adapter import check_and_adapt_order_details_if_necessary
 from octobot_trading.orders.order_factory import create_order_instance, create_order_instance_from_raw
 from octobot_trading.orders.order_util import get_pre_order_data
 from octobot_trading.trades.trade_factory import create_trade_from_order
