+++ setup.py
@@ -72,6 +72,7 @@ packages_list = ["octobot_trading.util.initializable",
                  "octobot_trading.data.portfolio",
                  "octobot_trading.data.portfolio_profitability",
                  "octobot_trading.data_adapters.candles_adapter",
+                 "octobot_trading.data_adapters.order_adapter",
                  "octobot_trading.data_manager.candles_manager",
                  "octobot_trading.data_manager.funding_manager",
                  "octobot_trading.data_manager.orders_manager",
@@ -84,7 +85,6 @@ packages_list = ["octobot_trading.util.initializable",
                  "octobot_trading.data_manager.order_book_manager",
                  "octobot_trading.data_manager.ticker_manager",
                  "octobot_trading.data_manager.recent_trades_manager",
-                 "octobot_trading.orders.order_adapter",
                  "octobot_trading.orders.order_factory",
                  "octobot_trading.orders.order_state",
                  "octobot_trading.orders.order_util",
