+++ tests/orders/test_order_adapter.py
@@ -17,7 +17,7 @@ import math
 import pytest
 
 from octobot_trading.enums import ExchangeConstantsMarketStatusColumns as Ecmsc
-from octobot_trading.orders.order_adapter import check_and_adapt_order_details_if_necessary, split_orders, \
+from octobot_trading.data_adapters.order_adapter import check_and_adapt_order_details_if_necessary, split_orders, \
     adapt_quantity, trunc_with_n_decimal_digits, add_dusts_to_quantity_if_necessary, adapt_price
 
 # All test coroutines will be treated as marked.
